move tests to test_pdb
